requirements.txt
flask==3.1.1
flask-JWT-Extended==4.7.1
flask-PyMongo==3.0.1
flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.1
pytest==8.2.1
pytest-flask==1.3.0
flask-cors==4.0.0